requirements.txt
chromadb==0.4.12
docx==0.2.4
docx2txt==0.8
exception==0.1.0
faiss-cpu==1.7.4
httptools==0.6.0
instructorembedding==1.0.1
langchain==0.0.137
langchainplus-sdk==0.0.20
langsmith==0.0.38
numexpr==2.8.6
openai==0.27.6
panel==1.2.3
pip-chill==1.0.3
pypdf==3.16.1
pypdf2==3.0.1
python-docx==0.8.11
python-dotenv==1.0.0
sentence-transformers==2.2.2
streamlit==1.18.1
tiktoken==0.4.0
unstructured==0.10.16
watchfiles==0.20.0
websockets==11.0.3
xlrd==2.0.1